requirements.txt
Django==1.10.3
Fabric==1.12.0
Pillow==3.4.2
djangorestframework==3.5.3
ecdsa==0.13
funcsigs==1.0.2
mock==2.0.0
paramiko==1.17.2
pbr==1.10.0
pycrypto==2.6.1
pysqlite==2.8.3
requests==2.11.1
requests-mock==1.1.0
six==1.10.0
wsgiref==0.1.2
